Pro Strings for Stacy's Mom
Should cover a good balance of all the guitar parts in the song.

Big differences between how you would play it for real:

Jumps to the main syncopated guitar stab in the middle of the main guitar riff, since the original 5-key guitar chart does it as well.

Bends during the guitar solo are charted to HOPOs since they're half-steps and nicely timed.